Finish sortable flex callbacks example, remove old example screen
